Resolve operand types when fetching the operands
The operand-types get resolved when the operands get fetched. The cpu stores an
accessible operand-type, which can be resolved to a core operand-type for
execution purposes at any time, which at this point doesn't inflict the operand
itself. When calling the resolve-function, the operand itself gets resolved,
and the stored accessible operand-type gets changed to the underlying core
operand-type, which prevents the resolution from happening twice if it's not
supposed to and from confusing the results.
Additionally, the code gets cleaned.

Signed-off-by: Paul Schuetz <pa.schuetz@web.de>